src/AIoT-Based Weather Forecasting System Using Machine Learning/python-web/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing and ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.5

# API and HTTP
httpx==0.25.1
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
